+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2020-05-13
+
+### Changed
+
+- Implementation of the bigquery `CREATE MODEL` syntax. Thanks [@barrywhart](https://github.com/barrywhart).
+- Bugfixes for:
+  - Edge cases for L006
+  - False alarms on L025
+  - `ORDER BY x NULLS FIRST|LAST`
+  - `FOR` keyword in bigquery `SYSTEM_TIME` syntax.
+
 ## [0.3.3] - 2020-05-11
 
 ### Added